extract the field annotation mapping and re-use in schema field
